refactor: split composables into header and footer
