+++ apps/web/app/components/blocks/Footer/Footer.vue
@@ -94,9 +94,7 @@ const props = defineProps<FooterProps>();
 const route = useRoute();
 const localePath = useLocalePath();
 const NuxtLink = resolveComponent('NuxtLink');
-const { footerBlock, mapFooterData, FOOTER_SWITCH_DEFINITIONS, createFooterBlock } = useBlockTemplates(
-  'index',
-  'immutable',
+const { footerBlock, mapFooterData, FOOTER_SWITCH_DEFINITIONS, createFooterBlock } = useFooterBlock(
   useNuxtApp().$i18n.locale.value,
 );
