+++ apps/web/app/components/blocks/Footer/FooterForm.vue
@@ -366,11 +366,12 @@ import { SfInput, SfTextarea, SfSwitch, SfIconCheck } from '@storefront-ui/vue';
 import type { FooterContent, FooterBlock } from './types';
 import type { Block } from '@plentymarkets/shop-api';
 const route = useRoute();
-const { data, mapFooterData, FOOTER_SWITCH_DEFINITIONS } = useBlockTemplates(
+const { data } = useBlockTemplates(
   route?.meta?.identifier as string,
   route.meta.type as string,
   useNuxtApp().$i18n.locale.value,
 );
+const { mapFooterData, FOOTER_SWITCH_DEFINITIONS } = useFooterBlock(useNuxtApp().$i18n.locale.value);
 const { blockUuid } = useSiteConfiguration();
 const { findOrDeleteBlockByUuid } = useBlockManager();
 const props = defineProps<{ uuid?: string }>();
